+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The released versions changelogs below are only about the main repository.
 - Anyone can now host a repository for its own CLI (edit cli_url option from the configuration file)
 - Repositories are compatibles with Github, but can also be simple directories listing web servers
 - Documentation website at https://bashpack-project.github.io
+- Subcommand template is available with the option "-n'
 
 ##### Modified
 - The CLI has been redesigned has a "script manager", to download/install/run/remove scripts as subcommands